Update intro.md
Our blockchain project is a paradigm shift in financial inclusion for Indian refugees. By seamlessly integrating decentralized identity management with a prepaid MasterCard, we offer a secure and transparent solution. Our innovation lies in the synergy of these components, creating a comprehensive ecosystem. The project not only empowers refugees economically but also contributes to a transparent and efficient financial landscape. This holistic approach, coupled with advanced blockchain technology, positions us as leaders in reshaping the narrative of financial inclusion.